R1.15)
In discussing these requirements with students from the music
department, they have indicated that they want to have song tempo
recorded with the data of the chart. Therefore modify the code so that
the XML file can specify a <tempo> 120 </tempo> tag. Tempo should be
displayed in much the same way as key, or musical style is displayed.